Split out extension points into SAWScript.Crucible.Common.Setup.Value
This is purely a refactor. This will be useful in a subsequent commit to avoid
import cycles.
